refactor: update imports dut to new file placement
