+++ app/src/main/kotlin/view/mainScreen/resetView.kt
@@ -8,7 +8,7 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import viewmodel.MainScreenViewModel
+import viewmodel.mainScreenVM.MainScreenViewModel
 
 @Composable
 fun <V> resetGraphView(viewModel: MainScreenViewModel<V>) {
