+++ app/src/main/kotlin/viewmodel/placementStrategy/RepresentationStrategy.kt
@@ -1,4 +1,6 @@
-package viewmodel
+package viewmodel.placementStrategy
+
+import viewmodel.graph.GraphViewModel
 
 interface RepresentationStrategy {
     fun <V> place(width: Double, height: Double,  graphViewModel: GraphViewModel<V>)
